+++ pilot/pkg/networking/deprecated/header.go
