+++ pilot/pkg/networking/deprecated/util.go
