+++ pilot/pkg/networking/interface.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package dataplane
+package networking
 
 import (
 	"github.com/envoyproxy/go-control-plane/envoy/api/v2"
